LayoutTests/fast/sub-pixel/auto-table-layout-should-avoid-text-wrapping.html

<!DOCTYPE html>
<html>
<head>
<style>
body {
  font-size: 11px;
  font-family: 'lucida grande',tahoma,verdana,arial,sans-serif;
  line-height: 20px;
}
td {
  padding: 0 0 0 0;
}
</style>
<script src="../../fast/js/resources/js-test-pre.js"></script>
</head>
<body>
<table id="testTable">
  <tbody>
    <tr>
      <td id="testCell">English (US)</td>
    </tr>
  </tbody>
</table>
<p id="description"></p>
<div id="console"></div>
<script>
description("This tests that we don't incorrectly wrap an auto-layout table due to improper truncation of sub-pixel accumulation. See Bug 93911.");

var lineHeight = document.getElementById('testCell').style.lineHeight;
shouldBe("document.getElementById('testCell').offsetHeight + 'px'", "getComputedStyle(document.getElementById('testCell')).lineHeight");

if (window.testRunner)
  document.getElementById("testTable").style.display = 'none';
</script>
<script src="../../fast/js/resources/js-test-post.js"></script>
</body>
