third_party/blink/web_tests/fast/sub-pixel/auto-table-layout-should-avoid-text-wrapping-expected.txt

This tests that we don't incorrectly wrap an auto-layout table due to improper truncation of sub-pixel accumulation. See Bug 93911.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.getElementById('testCell').offsetHeight + 'px' is getComputedStyle(document.getElementById('testCell')).lineHeight
PASS successfullyParsed is true

TEST COMPLETE
